+++ src/main/java/io/kontur/repository/CatCrudRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.Optional;
 import java.util.Set;
 
-public interface CatCrudRepository extends CrudRepository<Cat, Integer> {
+public interface CatCrudRepository extends CrudRepository<Cat, Long> {
   Optional<Cat> findCatByName(String name);
 }
