+++ src/main/java/io/kontur/repository/UserCrudRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface UserCrudRepository extends CrudRepository<User, Integer> {
+public interface UserCrudRepository extends CrudRepository<User, Long> {
   Optional<User> findUserByLogin(String name);
 }
